refactor: merge toolbar into titlebar, drop search/theme/settings buttons
Fold window controls into TitleBar and delete SharedToolbar. With search
and settings/theme buttons stripped from the old toolbar, the second row
held only the window controls plus an empty drag spacer — a redundant
bar whenever the tab strip was visible. Now TitleBar is the single top
chrome: tab strip when there are multiple tabs, drag region, and window
controls when enabled.

TitleBar renders when tabs > 1 OR showWindowControls is true, so users on
compositors that provide their own window chrome (e.g. Hyprland) can
hide it entirely by disabling showWindowControls.

The removed ThemeSwitcher dropdown and settings gear remain reachable:
ctrl+, still opens settings, and the command palette now lists a
"Theme: <Name>" entry per registered theme.
